refactor: 공지사항 도메인 패키지 이동
